Prepare latest app for Railway deployment
Promote the latest app surface, remove comments and duplicate helper files, and add the 21.6K scrap gold purity calculation.

+++ components/pricing-table.tsx
@@ -959,10 +959,10 @@ export function PricingTable({
       const getNumericValue = (purity: string): number => {
         if (purity.endsWith('K')) {
 
-          return parseInt(purity.replace('K', ''))
+          return parseFloat(purity.replace('K', ''))
         } else {
 
-          return parseInt(purity)
+          return parseFloat(purity)
         }
       }
       return getNumericValue(a) - getNumericValue(b)
